Refactor validation logic into separate files with improved error handling
