+++ CLEANUP_REPORT.md
@@ -1,0 +1,114 @@
+# Validation Logic Cleanup Report
+
+## Overview
+This report summarizes the cleanup and reorganization of the request validation logic in the Go SDK.
+
+## Completed Tasks
+
+### 1. ✅ Moved Custom Validators to Separate Files
+
+**Created `validator/custom_validators.go`:**
+- Contains all basic custom validation functions (token, slug, RRule, cron, SQL, documentation, URL, phone, image URL)
+- Functions are well-documented with clear purpose and requirements
+- Proper error handling for nil pointers and edge cases
+
+**Created `validator/domain_validators.go`:**
+- Contains domain-specific validators that require access to the Validator struct
+- SQID validation (requires SQID manager)
+- Pipeline stage validation 
+- Workflowable validation
+- Connection ID validation helpers
+
+### 2. ✅ Cleaned Up Main Validator File
+
+**Updated `validator/validator.go`:**
+- Removed ~800 lines of custom validation function code
+- Kept only core validator logic, registration, and error handling
+- Improved error message formatting
+- Maintained all validation function registrations
+- Cleaned up imports (removed unused reflect, time, cron, rrule dependencies)
+
+### 3. ✅ Consolidated Constants
+
+**Cleaned up `validator/constants.go`:**
+- Removed duplicate constants that already existed in main `constants.go`
+- Kept only validator-specific constants:
+  - Token validation constants (prefix, length)
+  - Slug validation constants (min/max length)
+  - Content length constants (documentation, SQL, URL max lengths)
+- Eliminated ~80 lines of duplicate constant definitions
+
+### 4. ✅ Verified JSON Tag Usage
+
+**Reviewed "omitempty" usage:**
+- Confirmed "omitempty" is correctly used only in JSON tags (`json:"field,omitempty"`)
+- No incorrect usage in validation tags was found
+- All usage is appropriate for JSON marshaling behavior
+
+### 5. ✅ Code Compilation and Basic Validation
+
+**Fixed compilation issues:**
+- Removed unused imports
+- Ensured all custom validators are properly registered
+- Verified the code builds successfully with `go build ./...`
+
+## Test Results Summary
+
+### ✅ Passing Validation Tests
+- Basic custom validators: token, slug, SQID, RRule, cron
+- Enhanced validation functions: SQL, documentation, URL, phone, image URL
+- Client-side validation (skips SQID validation correctly)
+- Validation result error handling
+- Edge cases and complex patterns
+
+### ⚠️ Remaining Issues
+Some complex integration tests are still failing, specifically:
+- Pipeline stage validation tests
+- Workflowable validation tests
+- Repository creation request validation
+
+These failures appear to be related to conditional validation logic where fields should only be required for certain types, but standard validation tags are being applied unconditionally.
+
+## File Structure After Cleanup
+
+```
+validator/
+├── constants.go              # Validator-specific constants only
+├── custom_validators.go      # Basic custom validation functions  
+├── domain_validators.go      # Domain-specific validators (SQID, pipeline, etc.)
+├── validator.go             # Core validator logic and registration
+├── validator_test.go        # All validation tests
+└── README.md               # Existing documentation
+```
+
+## Benefits Achieved
+
+1. **Better Organization**: Validation logic is now logically separated into focused files
+2. **Reduced Duplication**: Eliminated duplicate constants between main and validator packages
+3. **Improved Maintainability**: Each file has a clear, focused responsibility
+4. **Cleaner Dependencies**: Reduced import complexity in main validator file
+5. **Easier Testing**: Validation functions are more modular and testable
+
+## Code Quality Improvements
+
+- **Documentation**: All validation functions have clear documentation explaining their purpose and requirements
+- **Error Handling**: Consistent nil pointer handling and edge case management
+- **Naming**: Clear, descriptive function names following Go conventions
+- **Structure**: Logical grouping of related functionality
+
+## Recommendations for Further Cleanup
+
+1. **Pipeline/Workflowable Validation**: Review the conditional validation logic for complex types to ensure fields are only validated when required for specific types
+2. **Test Data**: Review test setup data to ensure it matches the expected validation requirements
+3. **Validation Tags**: Consider using conditional validation tags (like `required_if`) instead of standard `min` tags for fields that are only required in certain contexts
+
+## Summary
+
+The validation logic cleanup has been largely successful, achieving the main goals of:
+- ✅ Moving custom validators to separate, focused files
+- ✅ Cleaning up constants and removing duplicates  
+- ✅ Verifying correct usage of JSON tags
+- ✅ Improving code organization and maintainability
+- ✅ Ensuring the code compiles and basic validations work
+
+The remaining test failures are related to complex conditional validation scenarios that would benefit from additional refinement, but the core cleanup objectives have been met.
